+++ weldx/asdf/tags/weldx/core/data_array.py
@@ -1,6 +1,3 @@
-import numpy as np
-import pandas as pd
-from pandas.api.types import is_datetime64_any_dtype as is_datetime
 from xarray import DataArray
 
 import weldx.asdf.tags.weldx.core.common_types as ct
@@ -18,29 +15,66 @@ class XarrayDataArrayASDF(WeldxType):
 
     @classmethod
     def to_tree(cls, node: DataArray, ctx):
-        """Convert an xarray.DataArray to a tagged tree"""
+        """
+        Convert an instance of the 'xarray.DataArray' type into YAML  representations.
 
-        attributes = []
+        Parameters
+        ----------
+        node :
+            Instance of the 'xarray.DataArray' type to be serialized.
+
+        ctx :
+            An instance of the 'AsdfFile' object that is being written out.
+
+        Returns
+        -------
+            A basic YAML type ('dict', 'list', 'str', 'int', 'float', or
+            'complex') representing the properties of the 'xarray.DataArray'
+            type to be serialized.
+
+        """
+        attributes = node.attrs
         coordinates = []
-        variables = ct.Variable("data", node.dims, node.data)
+        data = ct.Variable("data", node.dims, node.data)
 
-        for name, data in node.coords.items():
-            coordinates.append(ct.Variable(name, data.dims, data.data))
+        for name, coord_data in node.coords.items():
+            coordinates.append(ct.Variable(name, coord_data.dims, coord_data.data))
 
         tree = {
             "attributes": attributes,
             "coordinates": coordinates,
-            "variables": variables,
+            "data": data,
         }
 
         return tree
 
     @classmethod
     def from_tree(cls, tree, ctx):
-        """Convert a tagged tree to an xarray.DataArray"""
-        data = tree["variables"].data
-        dims = tree["variables"].dimensions
+        """
+        Converts basic types representing YAML trees into an 'xarray.DataArray'.
+
+        Parameters
+        ----------
+        tree :
+            An instance of a basic Python type (possibly nested) that
+            corresponds to a YAML subtree.
+        ctx :
+            An instance of the 'AsdfFile' object that is being constructed.
+
+        Returns
+        -------
+        xarray.DataArray :
+            An instance of the 'xarray.DataArray' type.
+
+        """
+        data = tree["data"].data
+        dims = tree["data"].dimensions
         coords = {}
         for coordinate in tree["coordinates"]:
             coords[coordinate.name] = (coordinate.dimensions, coordinate.data)
-        return DataArray(data=data, coords=coords, dims=dims)
+
+        obj = DataArray(data=data, coords=coords, dims=dims)
+
+        obj.attrs = tree["attributes"]
+
+        return obj