+++ weldx/asdf/tags/weldx/core/dataset.py
@@ -15,8 +15,25 @@ class XarrayDatasetASDF(WeldxType):
 
     @classmethod
     def to_tree(cls, node: Dataset, ctx):
-        """Convert an xarray.Dataset to a tagged tree"""
-        attributes = []
+        """
+        Convert an instance of the 'xarray.Dataset' type into YAML representations.
+
+        Parameters
+        ----------
+        node :
+            Instance of the 'xarray.Dataset' type to be serialized.
+
+        ctx :
+            An instance of the 'AsdfFile' object that is being written out.
+
+        Returns
+        -------
+            A basic YAML type ('dict', 'list', 'str', 'int', 'float', or
+            'complex') representing the properties of the 'xarray.Dataset' type to be
+            serialized.
+
+        """
+        attributes = node.attrs
         coordinates = []
         dimensions = []
         variables = []
@@ -46,7 +63,23 @@ class XarrayDatasetASDF(WeldxType):
 
     @classmethod
     def from_tree(cls, tree, ctx):
-        """Convert a tagged tree to an xarray.Dataset"""
+        """
+        Converts basic types representing YAML trees into an 'xarray.Dataset'.
+
+        Parameters
+        ----------
+        tree :
+            An instance of a basic Python type (possibly nested) that
+            corresponds to a YAML subtree.
+        ctx :
+            An instance of the 'AsdfFile' object that is being constructed.
+
+        Returns
+        -------
+        xarray.Dataset :
+            An instance of the 'xarray.Dataset' type.
+
+        """
         data_vars = {}
 
         for variable in tree["variables"]:
@@ -56,4 +89,8 @@ class XarrayDatasetASDF(WeldxType):
         for coordinate in tree["coordinates"]:
             coords[coordinate.name] = (coordinate.dimensions, coordinate.data)
 
-        return Dataset(data_vars=data_vars, coords=coords)
+        obj = Dataset(data_vars=data_vars, coords=coords)
+
+        obj.attrs = tree["attributes"]
+
+        return obj