+++ CHANGELOG.md
@@ -7,3 +7,9 @@
 - add unit validation to `iso_groove-1.0.0.yaml` 
 - fixed const/enum constraints and properties in `iso_groove-1.0.0.yaml`
 - add some examples for testing `oneOf`, `anyOf` and other asdf keywords
+- add NetCDF inspired common types (`Dimension`,`Variable`) with corresponding
+ asdf serialization classes
+- add asdf serialization classes and schemas for `xarray.DataArray`, 
+`xarray.Dataset` and `weldx.transformations.LocalCoordinateSystem`.
+- add test for `xarray.DataArray`, `xarray.Dataset` and 
+`weldx.transformations.LocalCoordinateSystem` serialization.